# perf(web): fast path for utf8 decoding - part 1