- [x] Avoid copying buffers.

https://encoding.spec.whatwg.org/#dom-textdecoder-decode

> Implementations are strongly encouraged to use an implementation strategy that avoids this copy. When doing so they will have to make sure that changes to input do not affect future calls to [decode()](https://encoding.spec.whatwg.org/#dom-textdecoder-decode).

- [x] Special op to avoid string label deserialization and parsing. (Ideally we should map labels to integers in JS)
- [x] Avoid webidl `Object.assign` when options is undefined.

Upto ~1.8x improvement on small data.
```
# This patch
time 1893 ms rate 5282620
time 2067 ms rate 4837929
```

```
# main
time 3284 ms rate 3045066
time 3252 ms rate 3075030
```