@littledivy 2022-11-11
Revived by https://github.com/denoland/deno/pull/16593